% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_rejections2}
\alias{add_rejections2}
\title{Denote DE samples}
\usage{
add_rejections2(diff = dataDiff[[i]], alpha = 0.05, lfc = 1, sigCol = "p.adj")
}
\arguments{
\item{diff}{A list of SummarizedExperiment objects}

\item{alpha}{The significance cutoff level to use.}

\item{lfc}{The log2 fold change cutoff to use.}

\item{sigCol}{Either "p.val", "p.adj", or "BHCorrection" indicating the significance metric to use.}
}
\description{
Denote DE samples
}
