% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{AvailableGenes}
\alias{AvailableGenes}
\title{Select genes across features and datasets}
\usage{
AvailableGenes(SE, metaIndexDT, feature, datasets)

AvailableGenes(SE, metaIndexDT, feature, datasets)
}
\arguments{
\item{SE}{A sumarized experiment object containing expression level data.}

\item{metaIndexDT}{A data table of indexed meta data information.}

\item{feature}{The feature interested in exploring}

\item{datasets}{A vector of data sets to query.}
}
\description{
Select genes across features and datasets

Return available genes for specific data sets.
}
