% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{CreateMetaIndex}
\alias{CreateMetaIndex}
\title{Create an index file to reference meta data information}
\usage{
CreateMetaIndex(metaData)
}
\arguments{
\item{metaData}{A data table containing meta data information.}
}
\description{
Create an index file to reference meta data information
}
