% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DataFilter}
\alias{DataFilter}
\title{Data filtering functions}
\usage{
DataFilter(dataSeList, thr = 0)
}
\arguments{
\item{dataSeList}{A list of summarizedExperiment objects.}

\item{thr}{The number of missing values to filter by.}
}
\description{
Data filtering functions
}
