% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DataImpute}
\alias{DataImpute}
\title{Impute data}
\usage{
DataImpute(dataFiltList, type = "MinProb")
}
\arguments{
\item{dataFiltList}{A list of summarizedExperiment objects.}

\item{type}{Either "MinProb", "man", or "knn" indicating the type of imputation to perform.}
}
\description{
Impute data
}
