% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DEAnalysis}
\alias{DEAnalysis}
\title{Calculate differential expression}
\usage{
DEAnalysis(
  DataList = NormImpAll,
  type = "manual",
  ComparisonList = ComparisonList
)
}
\arguments{
\item{DataList}{A list of summarizedExperiment objects.}

\item{type}{Either "control", "all" or "manual" indicating the differential expression analysis algorithm to use.}
}
\description{
Calculate differential expression
}
