% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{densityPlotFromList}
\alias{densityPlotFromList}
\title{Density plots from a list}
\usage{
densityPlotFromList(MultiNormalizeList)
}
\arguments{
\item{MultiNormalizeList}{A list of summarizedExperiment objects.}
}
\description{
Density plots from a list
}
