% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DESEGenerate}
\alias{DESEGenerate}
\title{Create a summarized experiment object that integrates all DEG data.}
\usage{
DESEGenerate(DEGDatapath, SEPath)
}
\arguments{
\item{DEGDatapath}{The path to the directory where the fold change data are stored.}

\item{SEPath}{The path to the directory where the compiled SummarizedExperiment object should be stored.}
}
\description{
Create a summarized experiment object that integrates all DEG data.
}
