% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DESEProtGenerate}
\alias{DESEProtGenerate}
\title{Create a summarized experiment object that integrates all Proteomic DEG data.}
\usage{
DESEProtGenerate(ProtDatapath, SEPath)
}
\arguments{
\item{ProtDatapath}{The path to the directory where the fold change data are stored.}

\item{SEPath}{The path to the directory where the compiled SummarizedExperiment object should be stored.}
}
\description{
Create a summarized experiment object that integrates all Proteomic DEG data.
}
