% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DesignMatrixFromNames}
\alias{DesignMatrixFromNames}
\title{Design Matrix from file names}
\usage{
DesignMatrixFromNames(Fpath)
}
\arguments{
\item{Fpath}{The directory where the data is stored.}
}
\description{
Design Matrix from file names
}
