% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ExternalDataHarmonize}
\alias{ExternalDataHarmonize}
\title{Incorporate externally analyzed data into app.}
\usage{
ExternalDataHarmonize(Fpath, OutPath)
}
\arguments{
\item{Fpath}{The path to the directory where the analyzed data is stored.}

\item{OutPath}{The path to the directory where the formatted data is saved.}
}
\description{
Incorporate externally analyzed data into app.
}
