% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{FormatMaxQuant}
\alias{FormatMaxQuant}
\title{Format proteomics data. Note: due to the lack of consistency between data sets, each proteomics data set must be formatted individually. This function provides examples of how to format the data.}
\usage{
FormatMaxQuant(path)
}
\arguments{
\item{path}{The directory where the downloaded data is stored.}
}
\description{
Format proteomics data. Note: due to the lack of consistency between data sets, each proteomics data set must be formatted individually. This function provides examples of how to format the data.
}
