% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{GenerateRawSE}
\alias{GenerateRawSE}
\title{Generate raw data summarized experiment object.}
\usage{
GenerateRawSE(df, ArrayDT)
}
\arguments{
\item{df}{The meta data data frame created using the `MetDataCompile` function.}

\item{ArrayDT}{The Array raw data frame created using the `RawDataCompile` function}
}
\description{
Generate raw data summarized experiment object.
}
