% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{GEOCompile}
\alias{GEOCompile}
\title{GEO data download}
\usage{
GEOCompile(
  DS,
  gpl,
  gsm,
  namestr,
  nameraw,
  PlatAnnotInfo,
  destdir,
  filename = NULL,
  writeDB = TRUE,
  writeRaw = TRUE,
  GenerateMetaData,
  MetaDataPath,
  writeMetaData = TRUE,
  DBPath,
  Technology,
  renameRaw = FALSE,
  subsetRaw = FALSE,
  sleep = 30
)
}
\arguments{
\item{DS}{A vector containing the GEO GSE numbers of the desired datasets.}

\item{gpl}{A vector containing the GEO GPL numbers of the desired datasets.}

\item{gsm}{A vector containing the Comparison vector used in GEO2R describing how to parse the data and calculate fold changes.}

\item{namestr}{A vector containing the column names of the fold change calculations that are exported.}

\item{nameraw}{An optional vector containing the column names of the raw data columns that are exported.}

\item{PlatAnnotInfo}{The name of the object housing the plate annotation information exported from the `PlatformAnnotationLoad` function.}

\item{destdir}{The path to the directory to cache GEO data downloads.}

\item{filename}{description}

\item{writeDB}{Either TRUE or FALSE indicating if the fold change data should be saved to drive.}

\item{writeRaw}{A logical vector indicating if the raw data should be saved to drive.}

\item{GenerateMetaData}{A logical vector indicating if the meta data should be saved to drive.}

\item{MetaDataPath}{The path to the directory where the meta data should be saved.}

\item{writeMetaData}{Either TRUE or FALSE indicating if the meta data should be saved to drive.}

\item{DBPath}{The path to the directory where the raw and fold change data should be written.}

\item{Technology}{A character vector indicating either "Array" or "RNAseq" indicating if the data was generated using a minroArray or RNA sequencing technology.}

\item{renameRaw}{Either TRUE or FALSE indicating if the columns of the raw data should be renamed.}

\item{subsetRaw}{Either TRUE or FALSE indicating if the raw data should be subsetted to only the columns used in fold change calculations.}

\item{sleep}{The number of seconds each iteration should wait before starting the next.}
}
\description{
GEO data download
}
