% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{impute2}
\alias{impute2}
\title{Impute data}
\usage{
impute2(
  se,
  fun = c("bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "man", "min", "zero",
    "mixed", "nbavg"),
  ...
)
}
\arguments{
\item{se}{A list of SummarizedExperiment objects.}

\item{fun}{Either "MinProb", "man", or "knn" indicating the type of imputation to perform.}
}
\description{
Impute data
}
