% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{LowSampleCountRmove}
\alias{LowSampleCountRmove}
\title{Remove samples with a low count}
\usage{
LowSampleCountRmove(VarRMList, cut)
}
\arguments{
\item{VarRMList}{A list of summarizedExperiment objects.}

\item{cut}{The minimum cutoff for samples.}
}
\description{
Remove samples with a low count
}
