% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{makeDirectory}
\alias{makeDirectory}
\title{Create directories}
\usage{
makeDirectory(homedir)
}
\arguments{
\item{homedir}{The directory where the database will be stored}
}
\description{
Create directories
}
