% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MetDataCompile}
\alias{MetDataCompile}
\title{MetDataCompile}
\usage{
MetDataCompile(RNAseqFilePath, ArrayFilePath, overview)
}
\arguments{
\item{RNAseqFilePath}{The directory where the RNAseq metadata is saved.}

\item{ArrayFilePath}{The directory where the Array meta data is saved.}

\item{overview}{A data table containing the overview information.}
}
\description{
MetDataCompile
}
