% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{NPeptideThreshold}
\alias{NPeptideThreshold}
\title{N peptide filtering}
\usage{
NPeptideThreshold(dataFiltList, Npeptides = 2)
}
\arguments{
\item{dataFiltList}{A list of summarizedExperiment objects.}

\item{Npeptides}{The number of peptides to filter by.}
}
\description{
N peptide filtering
}
