% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{PlatformAnnotationLoad}
\alias{PlatformAnnotationLoad}
\title{Compile Array annotation information}
\usage{
PlatformAnnotationLoad(PlatInfo)
}
\arguments{
\item{PlatInfo}{A csv file containing a table of the Array information}
}
\description{
Compile Array annotation information
}
