% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_volcano2}
\alias{plot_volcano2}
\title{Volcano plot}
\usage{
plot_volcano2(
  dep,
  contrast,
  label_size = 3,
  add_names = TRUE,
  adjusted = FALSE,
  plot = TRUE,
  BHadjusted
)
}
\arguments{
\item{contrast}{A named vector of comparisons that were made.}

\item{label_size}{The size of the labels on the plot.}

\item{add_names}{A logical value indicating if data point names should be added to the plot.}

\item{adjusted}{A logical value indicating if adjusted significnce values should be plotted.}

\item{plot}{A logical value indicating if a plot should be returned.}

\item{BHadjusted}{A logical value indicating if the Benjamini–Hochberg method of adjusting p-values should be plotted.}

\item{DEP}{A SummarizedExperiment object}
}
\description{
Volcano plot
}
