% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{proteomicMZSave}
\alias{proteomicMZSave}
\title{Save formatted proteomics data}
\usage{
proteomicMZSave(MZList, path)
}
\arguments{
\item{MZList}{Formatted proteomics data stored as a list.}

\item{path}{The directory where the downloaded data is stored.}
}
\description{
Save formatted proteomics data
}
