% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ProteomicProteinName}
\alias{ProteomicProteinName}
\title{ProteomicProteinName}
\usage{
ProteomicProteinName(fPath)
}
\arguments{
\item{fPath}{The directory where the proteomic datasets are stored}
}
\description{
ProteomicProteinName
}
