% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ProteomicsDataDownload}
\alias{ProteomicsDataDownload}
\title{Return available genes for specific data sets.}
\usage{
ProteomicsDataDownload(path, DS, DownloadRawData = FALSE)
}
\arguments{
\item{path}{The directory where the downloaded data is stored.}

\item{DS}{The ProteomeExchange "PDX" numbers of the data to download.}

\item{DownloadRawData}{A logical value indicating if the raw data shoud be downloaded.}
}
\description{
Return available genes for specific data sets.
}
