% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ProteomicSELoad}
\alias{ProteomicSELoad}
\title{Load proteomics summarized experiment object from DB}
\usage{
ProteomicSELoad(Path = file.path(homedir, "Proteomic_3"), Fname = files[1])
}
\arguments{
\item{Path}{The path to where the data should be saved.}

\item{Fname}{The file name of the dataset to load.}
}
\description{
Load proteomics summarized experiment object from DB
}
