% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ProtSELoad}
\alias{ProtSELoad}
\title{Create summarized experiment objects}
\usage{
ProtSELoad(DesignDT, Fpath)
}
\arguments{
\item{DesignDT}{A design table containing the dataset, column label, and the treatment condition for samples.}

\item{Fpath}{The directory where the data is stored.}
}
\description{
Create summarized experiment objects
}
