% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{RawDataCompile}
\alias{RawDataCompile}
\title{Combine all raw data into one data table.}
\usage{
RawDataCompile(
  Fpath = file.path(homedir, "5_AppData"),
  outPath = file.path("./6_ProcessFiles/RawData.txt"),
  StartAt = 1,
  sleep = 20,
  DataType = "Array",
  CheckIdenticalData = FALSE
)
}
\arguments{
\item{Fpath}{The file path to the directory where the raw data is stored.}

\item{outPath}{The path to the assembled raw data file.}

\item{StartAt}{A numeric value indicating the position of the raw datasets to start compiling raw data from.}

\item{sleep}{The number of seconds each iteration should wait before starting the next.}

\item{DataType}{Either "Array", "RNAseqRPKM", or "RNAseqCount" indicating the type of raw data to assemble into a data frame.  Only "Array" is used for data generated using GEO2R.}

\item{SampleSelect}{A logical vector indicating which raw datasets to incorporate into the app.}
}
\description{
Combine all raw data into one data table.
}
