% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{RowDataCompile}
\alias{RowDataCompile}
\title{Obtain row bound data}
\usage{
RowDataCompile(tissueSplitList)
}
\arguments{
\item{tissueSplitList}{A list of summarizedExperiment objects.}
}
\description{
Obtain row bound data
}
