% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SaveToProteomicDB}
\alias{SaveToProteomicDB}
\title{Save DEP expression summarizedExperiment objects}
\usage{
SaveToProteomicDB(SEList, Path)
}
\arguments{
\item{SEList}{A list of summarizedExperiment objects.}

\item{Path}{The path to where the data should be saved.}
}
\description{
Save DEP expression summarizedExperiment objects
}
