% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SigDEAnnotate}
\alias{SigDEAnnotate}
\title{Denote DE samples}
\usage{
SigDEAnnotate(dataDiff, alpha = 0.05, lfc = log2(2), sigCol = "p.adj")
}
\arguments{
\item{dataDiff}{A list of SummarizedExperiment objects}

\item{alpha}{The significance cutoff level to use.}

\item{lfc}{The log2 fold change cutoff to use.}

\item{sigCol}{Either "p.val", "p.adj", or "BHCorrection" indicating the significance metric to use.}
}
\description{
Denote DE samples
}
