% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano plot}
\usage{
VolcanoPlot(
  DEPList = DEPList,
  ComparisonList = ComparisonList,
  ymin = 0,
  ymax = 15,
  xmin = -3,
  xmax = 3,
  addNames = TRUE,
  Adjusted = TRUE,
  BHadjust
)
}
\arguments{
\item{DEPList}{A list of SummarizedExperiment objects}

\item{ComparisonList}{A named vector of comparisons that were made.}

\item{ymin}{The minimum y axis value.}

\item{ymax}{The maximum y axis value.}

\item{xmin}{The minimum x axis value.}

\item{xmax}{The maximum x axis value.}

\item{addNames}{A logical value indicating if data point names should be added to the plot.}

\item{Adjusted}{A logical value indicating if adjusted significnce values should be plotted.}

\item{BHadjust}{A logical value indicating if the Benjamini–Hochberg method of adjusting p-values should be plotted.}
}
\description{
Volcano plot
}
