% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{GenerateRawSE}
\alias{GenerateRawSE}
\title{GenerateRawSE}
\usage{
GenerateRawSE(df, ArrayDT, overview)
}
\arguments{
\item{df}{A data frame containing meta data.}

\item{ArrayDT}{A data table containing compiled raw data.}
}
\description{
GenerateRawSE
}
