% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MetDataCompile}
\alias{MetDataCompile}
\title{Compile all meta data into one table}
\usage{
MetDataCompile(RNAseqFilePath, ArrayFilePath, overview = overview)
}
\arguments{
\item{RNAseqFilePath}{The file path where the RNAseq meta data is stored.  Files should be named according to the following naming convention: "GSEXXXXX_SraRunTable.txt"}

\item{ArrayFilePath}{The file path where the Array meta data is stored.}
}
\description{
Compile all meta data into one table
}
