% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MultiNormalization}
\alias{MultiNormalization}
\title{Multiple normalization}
\usage{
MultiNormalization(dataPepCutOff)
}
\arguments{
\item{dataPepCutOff}{A list of SummarizedExperiment objects.}
}
\description{
Multiple normalization
}
