% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{RawDataCompile}
\alias{RawDataCompile}
\title{RawDataCompile}
\usage{
RawDataCompile(
  Fpath = file.path(homedir, "AppData"),
  outPath = file.path("./6_ProcessFiles/RawData.txt"),
  StartAt = 1,
  sleep = 20,
  DataType = "Array"
)
}
\arguments{
\item{Fpath}{The directory where raw tables are stored.}

\item{outPath}{The directory where the compiled table should be saved}

\item{StartAt}{Indicates the file number to start combining data from.}

\item{sleep}{The amount of time to pause between incorporating the next dataset. Only needed if using a machine that syncronizes the file to the cloud.}

\item{DataType}{Either "Array", "RNAseqRPKM", or "RNAseqCount" indicating the type of data that is incorporated.}
}
\description{
RawDataCompile
}
